refactor(transform): convert transform to a project
