+++ transform/README.md
@@ -1,0 +1,5 @@
+# Transform
+
+Tools to execute data transformations.
+
+This is necessary to make raw TSVs returned by Hue indexable by Solr.
